+++ README.ja.md
@@ -59,20 +59,26 @@ $ go test -cover ./...
     "404_img_path": "/path/to/404/image",
     "access_log_path": "/path/to/access/log",
     "error_log_path": "/path/to/error/log",
-    "providers": {
-        "alias/0" : {
-            "type" : "web",
-            "src" : "http://aaa.com/bbb"
+    "providers": [
+        {
+            "alias/0" : {
+                "type" : "web",
+                "src" : "http://aaa.com/bbb"
+            }
         },
-        "alias/1" : {
-            "type" : "web",
-            "src" : "https://ccc.com"
+        {
+            "alias/1" : {
+                "type" : "web",
+                "src" : "https://ccc.com"
+            }
         },
-        "alias/3" : {
-            "type" : "s3",
-            "src" : "s3://bucket/path",
-            "region" : "ap-northeast-1"
+        {
+            "alias/3" : {
+                "type" : "s3",
+                "src" : "s3://bucket/path",
+                "region" : "ap-northeast-1"
+            }
         }
-    }
+    ]
 }
 ```
